Db migration

Approved-by: Зеленкевич Илья

+++ src/main/java/app/utils/processor/impl/product/ProductDeleteProcessor.java
@@ -1,14 +1,16 @@
 package app.utils.processor.impl.product;
 
-import app.service.ProductService;
-import app.service.impl.ProductServiceImpl;
+import app.service.impl.ProductService;
 import app.utils.SecurityContextHolder;
 import app.utils.processor.Processor;
+import lombok.extern.log4j.Log4j2;
 
 import static app.entity.user.enums.Role.ADMIN;
+import static app.utils.SecurityContextHolder.isCurrentUserSignedIn;
 
+@Log4j2
 public class ProductDeleteProcessor implements Processor {
-    private final ProductService productService = new ProductServiceImpl();
+    private final ProductService productService = new ProductService();
 
     @Override
     public boolean supports(String command) {
@@ -17,7 +19,7 @@ public class ProductDeleteProcessor implements Processor {
 
     @Override
     public void process(String command) {
-        if (isAdmin()) {
+        if (isCurrentUserSignedIn() && isAdmin()) {
             final Long id = Long.parseLong(command.split(" ")[1]);
 
             productService.deleteById(id);
